+++ packages/cloud_firestore/cloud_firestore_web/lib/cloud_firestore_web.dart
@@ -108,6 +108,11 @@ class FirebaseFirestoreWeb extends FirebaseFirestorePlatform {
   }
 
   @override
+  Settings get settings {
+    return const Settings();
+  }
+
+  @override
   set settings(Settings settings) {
     int? cacheSizeBytes;