+++ packages/firebase_database/firebase_database/lib/firebase_database.dart
@@ -8,10 +8,12 @@ library firebase_database;
 import 'dart:async';
 
 import 'package:firebase_core/firebase_core.dart';
+import 'package:firebase_database_platform_interface/firebase_database_platform_interface.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 
-import 'src/utils/push_id_generator.dart';
+export 'package:firebase_database_platform_interface/firebase_database_platform_interface.dart'
+    show ServerValue, MutableData, TransactionHandler, EventType;
 
 part 'src/database_reference.dart';
 part 'src/event.dart';