+++ packages/firebase_messaging/firebase_messaging/example/lib/permissions.dart
@@ -1,5 +1,4 @@
 // ignore_for_file: require_trailing_commas
-// @dart=2.9
 
 import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:flutter/foundation.dart';
@@ -14,7 +13,7 @@ class Permissions extends StatefulWidget {
 class _Permissions extends State<Permissions> {
   bool _requested = false;
   bool _fetching = false;
-  NotificationSettings _settings;
+  late NotificationSettings _settings;
 
   Future<void> requestPermissions() async {
     setState(() {
@@ -61,16 +60,16 @@ class _Permissions extends State<Permissions> {
     }
 
     return Column(children: [
-      row('Authorization Status', statusMap[_settings.authorizationStatus]),
+      row('Authorization Status', statusMap[_settings.authorizationStatus]!),
       if (defaultTargetPlatform == TargetPlatform.iOS) ...[
-        row('Alert', settingsMap[_settings.alert]),
-        row('Announcement', settingsMap[_settings.announcement]),
-        row('Badge', settingsMap[_settings.badge]),
-        row('Car Play', settingsMap[_settings.carPlay]),
-        row('Lock Screen', settingsMap[_settings.lockScreen]),
-        row('Notification Center', settingsMap[_settings.notificationCenter]),
-        row('Show Previews', previewMap[_settings.showPreviews]),
-        row('Sound', settingsMap[_settings.sound]),
+        row('Alert', settingsMap[_settings.alert]!),
+        row('Announcement', settingsMap[_settings.announcement]!),
+        row('Badge', settingsMap[_settings.badge]!),
+        row('Car Play', settingsMap[_settings.carPlay]!),
+        row('Lock Screen', settingsMap[_settings.lockScreen]!),
+        row('Notification Center', settingsMap[_settings.notificationCenter]!),
+        row('Show Previews', previewMap[_settings.showPreviews]!),
+        row('Sound', settingsMap[_settings.sound]!),
       ],
       ElevatedButton(
           onPressed: () => {}, child: const Text('Reload Permissions')),