+++ packages/firebase_performance/firebase_performance_platform_interface/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.0.1+3
+
+ - **FIX**: fix the token verification. (#6912).
+
+## 0.0.1+2
+
+  **FIX**: fix the token verification (#6912).
+
+
 ## 0.0.1+1
 
  - **REFACTOR**: create interface plugin that federates the main plugin (#6851).